first pass at live loan editing, saving changes to sql item, and reloading UI to reflect changes

+++ app/src/main/java/com/prototype/balcorasystems/slw3/Object_Loan.java
@@ -45,7 +45,7 @@ public class Object_Loan {
         return inceptionDate;
     }
 
-    public void setInceptionDate(int inceptionDate) {
+    public void setInceptionDate(long inceptionDate) {
         this.inceptionDate = inceptionDate;
     }
